@@ -1,23 +1,23 @@
-const qiniu = require("../index.js");
-const proc = require("process");
+const qiniu = require('../index.js');
+const proc = require('process');
 
 var bucket = proc.env.QINIU_TEST_BUCKET;
 var accessKey = proc.env.QINIU_ACCESS_KEY;
 var secretKey = proc.env.QINIU_SECRET_KEY;
 var mac = new qiniu.auth.digest.Mac(accessKey, secretKey);
 var options = {
-  scope: bucket,
-}
+    scope: bucket
+};
 var putPolicy = new qiniu.rs.PutPolicy(options);
 
 var uploadToken = putPolicy.uploadToken(mac);
 var config = new qiniu.conf.Config();
-var localFile = "/Users/jemy/Downloads/VIDEO_20191008_093955.mp4.zip";
-//config.zone = qiniu.zone.Zone_z0;
+var localFile = '/Users/jemy/Downloads/VIDEO_20191008_093955.mp4.zip';
+// config.zone = qiniu.zone.Zone_z0;
 var formUploader = new qiniu.form_up.FormUploader(config);
 var putExtra = new qiniu.form_up.PutExtra();
 
-//bytes
+// bytes
 // formUploader.put(uploadToken, null, "hello", null, function(respErr,
 //   respBody, respInfo) {
 //   if (respErr) {
@@ -32,37 +32,37 @@ var putExtra = new qiniu.form_up.PutExtra();
 //   }
 // });
 
-//file
+// file
 putExtra.fname = 'testfile16.mp4';
-formUploader.putFile(uploadToken, null, localFile, putExtra, function(respErr,
-  respBody, respInfo) {
-  if (respErr) {
-    throw respErr;
-  }
-
-  if (respInfo.statusCode == 200) {
-    console.log(respBody);
-  } else {
-    console.log(respInfo.statusCode);
-    console.log(respBody);
-  }
-});
-
-//test query zone frequently when config.zone==null;
-setTimeout(upload,1500);
-function upload (){
-    putExtra.fname = 'testfile17.mp4';
-    formUploader.putFile(uploadToken, null, localFile, putExtra, function(respErr,
-      respBody, respInfo) {
-      if (respErr) {
+formUploader.putFile(uploadToken, null, localFile, putExtra, function (respErr,
+    respBody, respInfo) {
+    if (respErr) {
         throw respErr;
-      }
+    }
 
-      if (respInfo.statusCode == 200) {
+    if (respInfo.statusCode == 200) {
         console.log(respBody);
-      } else {
+    } else {
         console.log(respInfo.statusCode);
         console.log(respBody);
-      }
+    }
+});
+
+// test query zone frequently when config.zone==null;
+setTimeout(upload, 1500);
+function upload () {
+    putExtra.fname = 'testfile17.mp4';
+    formUploader.putFile(uploadToken, null, localFile, putExtra, function (respErr,
+        respBody, respInfo) {
+        if (respErr) {
+            throw respErr;
+        }
+
+        if (respInfo.statusCode == 200) {
+            console.log(respBody);
+        } else {
+            console.log(respInfo.statusCode);
+            console.log(respBody);
+        }
     });
 }
