@@ -2,7 +2,7 @@ var conf = require('../conf');
 
 exports.Mac = Mac;
 
-function Mac(accessKey, secretKey) {
+function Mac (accessKey, secretKey) {
     this.accessKey = accessKey || conf.ACCESS_KEY;
     this.secretKey = secretKey || conf.SECRET_KEY;
 }
