@@ -4,39 +4,39 @@ const pkg = require('../package.json');
 exports.ACCESS_KEY = '<PLEASE APPLY YOUR ACCESS KEY>';
 exports.SECRET_KEY = '<DONT SEND YOUR SECRET KEY TO ANYONE>';
 
-var defaultUserAgent = function() {
+var defaultUserAgent = function () {
     return 'QiniuNodejs/' + pkg.version + ' (' + os.type() + '; ' + os.platform() +
     '; ' + os.arch() + '; )';
 };
 
 exports.USER_AGENT = defaultUserAgent();
-exports.BLOCK_SIZE = 4 * 1024 * 1024; //4MB, never change
+exports.BLOCK_SIZE = 4 * 1024 * 1024; // 4MB, never change
 
-//define api form mime type
+// define api form mime type
 exports.FormMimeUrl = 'application/x-www-form-urlencoded';
 exports.FormMimeJson = 'application/json';
 exports.FormMimeRaw = 'application/octet-stream';
 exports.RS_HOST = 'http://rs.qiniu.com';
-exports.RPC_TIMEOUT = 120000; //120s
+exports.RPC_TIMEOUT = 120000; // 120s
 exports.UC_HOST = 'uc.qbox.me';
 
-//proxy
+// proxy
 exports.RPC_HTTP_AGENT = null;
 exports.RPC_HTTPS_AGENT = null;
 
-exports.Config = function Config(options) {
+exports.Config = function Config (options) {
     options = options || {};
-    //use http or https protocol
+    // use http or https protocol
     this.useHttpsDomain = !!(options.useHttpsDomain || false);
-    //use cdn accerlated domains
+    // use cdn accerlated domains
     this.useCdnDomain = !!(options.useCdnDomain && true);
-    //zone of the bucket
-    //z0 huadong, z1 huabei, z2 huanan, na0 beimei
+    // zone of the bucket
+    // z0 huadong, z1 huabei, z2 huanan, na0 beimei
     this.zone = options.zone || null;
     this.zoneExpire = options.zoneExpire || -1;
 };
 
-exports.Zone = function(srcUpHosts, cdnUpHosts, ioHost, rsHost, rsfHost,
+exports.Zone = function (srcUpHosts, cdnUpHosts, ioHost, rsHost, rsfHost,
     apiHost) {
     this.srcUpHosts = srcUpHosts || {};
     this.cdnUpHosts = cdnUpHosts || {};
