@@ -1,18 +1,18 @@
 var util = require('./util');
 
-function Credentials(accessKey, secretKey) {
+function Credentials (accessKey, secretKey) {
     this.accessKey = accessKey;
     this.secretKey = secretKey;
 }
 
-Credentials.prototype.generateAccessToken = function(options, data) {
+Credentials.prototype.generateAccessToken = function (options, data) {
     var sign = this._signRequest(options, data);
     var token = 'Qiniu' + ' ' + this.accessKey + ':' + sign;
 
     return token;
 };
 
-Credentials.prototype._signRequest = function(options, body) {
+Credentials.prototype._signRequest = function (options, body) {
     var contentType = options.headers['Content-Type'];
 
     var host = options.host;
@@ -31,24 +31,20 @@ Credentials.prototype._signRequest = function(options, body) {
         data += body;
     }
 
-
     var digest = util.hmacSha1(data, this.secretKey);
-
 
     var sageDigest = util.base64ToUrlSafe(digest);
 
     return sageDigest;
 };
 
-Credentials.prototype.sign = function(data) {
+Credentials.prototype.sign = function (data) {
     var digest = util.hmacSha1(data, this.secretKey);
     var sageDigest = util.base64ToUrlSafe(digest);
     return this.accessKey + ':' + sageDigest;
-
 };
 
-Credentials.prototype.signJson = function(opt) {
-
+Credentials.prototype.signJson = function (opt) {
     var str = JSON.stringify(opt);
     var encodedStr = util.urlsafeBase64Encode(str);
     var sign = util.hmacSha1(encodedStr, this.secretKey);
