@@ -11,7 +11,7 @@ const BlockStream = require('block-stream2');
 exports.ResumeUploader = ResumeUploader;
 exports.PutExtra = PutExtra;
 
-function ResumeUploader(config) {
+function ResumeUploader (config) {
     this.config = config || new conf.Config();
 }
 
@@ -21,7 +21,7 @@ function ResumeUploader(config) {
 // @param mimeType                    指定文件的mimeType
 // @param resumeRecordFile            断点续传的已上传的部分信息记录文件
 // @param progressCallback(uploadBytes, totalBytes) 上传进度回调
-function PutExtra(fname, params, mimeType, resumeRecordFile, progressCallback) {
+function PutExtra (fname, params, mimeType, resumeRecordFile, progressCallback) {
     this.fname = fname || '';
     this.params = params || {};
     this.mimeType = mimeType || null;
@@ -29,7 +29,7 @@ function PutExtra(fname, params, mimeType, resumeRecordFile, progressCallback) {
     this.progressCallback = progressCallback || null;
 }
 
-ResumeUploader.prototype.putStream = function(uploadToken, key, rsStream,
+ResumeUploader.prototype.putStream = function (uploadToken, key, rsStream,
     rsStreamLen, putExtra, callbackFunc) {
     putExtra = putExtra || new PutExtra();
     if (!putExtra.mimeType) {
@@ -37,19 +37,18 @@ ResumeUploader.prototype.putStream = function(uploadToken, key, rsStream,
     }
 
     if (!putExtra.fname) {
-        putExtra.fname = key ? key : '?';
+        putExtra.fname = key || '?';
     }
 
-    rsStream.on('error', function(err) {
-    //callbackFunc
+    rsStream.on('error', function (err) {
+    // callbackFunc
         callbackFunc(err, null, null);
         rsStream.close();
-        return;
     });
 
     var useCache = false;
     var that = this;
-    if (this.config.zone!=""&&this.config.zone!=null) {
+    if (this.config.zone != '' && this.config.zone != null) {
         if (this.config.zoneExpire == -1) {
             useCache = true;
         } else {
@@ -65,7 +64,7 @@ ResumeUploader.prototype.putStream = function(uploadToken, key, rsStream,
         putReq(this.config, uploadToken, key, rsStream, rsStreamLen, putExtra,
             callbackFunc);
     } else {
-        zone.getZoneInfo(accessKey, bucket, function(err, cZoneInfo,
+        zone.getZoneInfo(accessKey, bucket, function (err, cZoneInfo,
             cZoneExpire) {
             if (err) {
                 callbackFunc(err, null, null);
@@ -73,12 +72,12 @@ ResumeUploader.prototype.putStream = function(uploadToken, key, rsStream,
                 return;
             }
 
-            //update object
+            // update object
             that.config.zone = cZoneInfo;
-            that.config.zoneExpire = cZoneExpire+parseInt(Date.now() / 1000);
+            that.config.zoneExpire = cZoneExpire + parseInt(Date.now() / 1000);
             this.config = that.config;
 
-            //req
+            // req
             putReq(that.config, uploadToken, key, rsStream, rsStreamLen,
                 putExtra,
                 callbackFunc);
@@ -86,25 +85,25 @@ ResumeUploader.prototype.putStream = function(uploadToken, key, rsStream,
     }
 };
 
-function putReq(config, uploadToken, key, rsStream, rsStreamLen, putExtra,
+function putReq (config, uploadToken, key, rsStream, rsStreamLen, putExtra,
     callbackFunc) {
-    //set up hosts order
+    // set up hosts order
     var upHosts = [];
 
     if (config.useCdnDomain) {
         if (config.zone.cdnUpHosts) {
-            config.zone.cdnUpHosts.forEach(function(host) {
+            config.zone.cdnUpHosts.forEach(function (host) {
                 upHosts.push(host);
             });
         }
-        config.zone.srcUpHosts.forEach(function(host) {
+        config.zone.srcUpHosts.forEach(function (host) {
             upHosts.push(host);
         });
     } else {
-        config.zone.srcUpHosts.forEach(function(host) {
+        config.zone.srcUpHosts.forEach(function (host) {
             upHosts.push(host);
         });
-        config.zone.cdnUpHosts.forEach(function(host) {
+        config.zone.cdnUpHosts.forEach(function (host) {
             upHosts.push(host);
         });
     }
@@ -122,20 +121,20 @@ function putReq(config, uploadToken, key, rsStream, rsStreamLen, putExtra,
     var finishedBlock = 0;
     var finishedCtxList = [];
     var finishedBlkPutRets = [];
-    //read resumeRecordFile
+    // read resumeRecordFile
     if (putExtra.resumeRecordFile) {
         try {
             var resumeRecords = fs.readFileSync(putExtra.resumeRecordFile).toString();
             var blkputRets = JSON.parse(resumeRecords);
 
             for (var index = 0; index < blkputRets.length; index++) {
-                //check ctx expired or not
+                // check ctx expired or not
                 var blkputRet = blkputRets[index];
                 var expiredAt = blkputRet.expired_at;
-                //make sure the ctx at least has one day expiration
+                // make sure the ctx at least has one day expiration
                 expiredAt += 3600 * 24;
                 if (util.isTimestampExpired(expiredAt)) {
-                    //discard these ctxs
+                    // discard these ctxs
                     break;
                 }
 
@@ -147,20 +146,19 @@ function putReq(config, uploadToken, key, rsStream, rsStreamLen, putExtra,
         }
     }
 
-    //check when to mkblk
-    blkStream.on('data', function(chunk) {
+    // check when to mkblk
+    blkStream.on('data', function (chunk) {
         readLen += chunk.length;
-        curBlock += 1; //set current block
+        curBlock += 1; // set current block
         if (curBlock > finishedBlock) {
             blkStream.pause();
-            mkblkReq(upDomain, uploadToken, chunk, function(respErr,
+            mkblkReq(upDomain, uploadToken, chunk, function (respErr,
                 respBody,
                 respInfo) {
                 var bodyCrc32 = parseInt('0x' + getCrc32(chunk));
                 if (respInfo.statusCode != 200 || respBody.crc32 != bodyCrc32) {
                     callbackFunc(respErr, respBody, respInfo);
                     rsStream.close();
-                    return;
                 } else {
                     finishedBlock += 1;
                     var blkputRet = respBody;
@@ -188,20 +186,20 @@ function putReq(config, uploadToken, key, rsStream, rsStreamLen, putExtra,
     });
 }
 
-function mkblkReq(upDomain, uploadToken, blkData, callbackFunc) {
-    //console.log("mkblk");
+function mkblkReq (upDomain, uploadToken, blkData, callbackFunc) {
+    // console.log("mkblk");
     var requestURI = upDomain + '/mkblk/' + blkData.length;
     var auth = 'UpToken ' + uploadToken;
     var headers = {
-        'Authorization': auth,
+        Authorization: auth,
         'Content-Type': 'application/octet-stream'
     };
     rpc.post(requestURI, blkData, headers, callbackFunc);
 }
 
-function mkfileReq(upDomain, uploadToken, fileSize, ctxList, key, putExtra,
+function mkfileReq (upDomain, uploadToken, fileSize, ctxList, key, putExtra,
     callbackFunc) {
-    //console.log("mkfile");
+    // console.log("mkfile");
     var requestURI = upDomain + '/mkfile/' + fileSize;
     if (key) {
         requestURI += '/key/' + util.urlsafeBase64Encode(key);
@@ -213,7 +211,7 @@ function mkfileReq(upDomain, uploadToken, fileSize, ctxList, key, putExtra,
         requestURI += '/fname/' + util.urlsafeBase64Encode(putExtra.fname);
     }
     if (putExtra.params) {
-    //putExtra params
+    // putExtra params
         for (var k in putExtra.params) {
             if (k.startsWith('x:') && putExtra.params[k]) {
                 requestURI += '/' + k + '/' + util.urlsafeBase64Encode(putExtra.params[
@@ -223,11 +221,11 @@ function mkfileReq(upDomain, uploadToken, fileSize, ctxList, key, putExtra,
     }
     var auth = 'UpToken ' + uploadToken;
     var headers = {
-        'Authorization': auth,
+        Authorization: auth,
         'Content-Type': 'application/octet-stream'
     };
     var postBody = ctxList.join(',');
-    rpc.post(requestURI, postBody, headers, function(err, ret, info) {
+    rpc.post(requestURI, postBody, headers, function (err, ret, info) {
         if (info.statusCode == 200 || info.statusCode == 701 ||
       info.statusCode == 401) {
             if (putExtra.resumeRecordFile) {
@@ -238,11 +236,11 @@ function mkfileReq(upDomain, uploadToken, fileSize, ctxList, key, putExtra,
     });
 }
 
-ResumeUploader.prototype.putFile = function(uploadToken, key, localFile,
+ResumeUploader.prototype.putFile = function (uploadToken, key, localFile,
     putExtra, callbackFunc) {
     putExtra = putExtra || new PutExtra();
-    var rsStream = fs.createReadStream(localFile,{
-        highWaterMark: conf.BLOCK_SIZE,
+    var rsStream = fs.createReadStream(localFile, {
+        highWaterMark: conf.BLOCK_SIZE
     });
     var rsStreamLen = fs.statSync(localFile).size;
     if (!putExtra.mimeType) {
@@ -257,7 +255,7 @@ ResumeUploader.prototype.putFile = function(uploadToken, key, localFile,
         callbackFunc);
 };
 
-ResumeUploader.prototype.putFileWithoutKey = function(uploadToken, localFile,
+ResumeUploader.prototype.putFileWithoutKey = function (uploadToken, localFile,
     putExtra, callbackFunc) {
     return this.putFile(uploadToken, null, localFile, putExtra, callbackFunc);
 };
