@@ -7,44 +7,42 @@ const fs = require('fs');
 const console = require('console');
 
 // eslint-disable-next-line no-undef
-before(function(done) {
-    if (!process.env.QINIU_ACCESS_KEY || !process.env.QINIU_SECRET_KEY || !
-    process.env.QINIU_TEST_BUCKET || !process.env.QINIU_TEST_DOMAIN) {
+before(function (done) {
+    if (!process.env.QINIU_ACCESS_KEY || !process.env.QINIU_SECRET_KEY || !process.env.QINIU_TEST_BUCKET || !process.env.QINIU_TEST_DOMAIN) {
         console.log('should run command `source test-env.sh` first\n');
         process.exit(0);
     }
     done();
 });
 
-
-//file to upload
+// file to upload
 var testFilePath1 = path.join(__dirname, 'logo.png');
 var testFilePath2 = path.join(__dirname, 'github.png');
 
 // eslint-disable-next-line no-undef
-describe('test form up', function() {
+describe('test form up', function () {
     var accessKey = proc.env.QINIU_ACCESS_KEY;
     var secretKey = proc.env.QINIU_SECRET_KEY;
     var bucket = proc.env.QINIU_TEST_BUCKET;
     var mac = new qiniu.auth.digest.Mac(accessKey, secretKey);
     var config = new qiniu.conf.Config();
-    //config.useHttpsDomain = true;
+    // config.useHttpsDomain = true;
     config.zone = qiniu.zone.Zone_z0;
     var bucketManager = new qiniu.rs.BucketManager(mac, config);
 
-    //delete all the files uploaded
+    // delete all the files uploaded
     var keysToDelete = [];
 
     // eslint-disable-next-line no-undef
-    after(function(done) {
+    after(function (done) {
         var deleteOps = [];
-        keysToDelete.forEach(function(key) {
+        keysToDelete.forEach(function (key) {
             deleteOps.push(qiniu.rs.deleteOp(bucket, key));
         });
 
-        bucketManager.batch(deleteOps, function(respErr, respBody) {
-            //console.log(respBody);
-            respBody.forEach(function(ret) {
+        bucketManager.batch(deleteOps, function (respErr, respBody) {
+            // console.log(respBody);
+            respBody.forEach(function (ret) {
                 ret.should.eql({
                     code: 200
                 });
@@ -54,7 +52,7 @@ describe('test form up', function() {
     });
 
     var options = {
-        scope: bucket,
+        scope: bucket
     };
     var putPolicy = new qiniu.rs.PutPolicy(options);
     var uploadToken = putPolicy.uploadToken(mac);
@@ -63,13 +61,13 @@ describe('test form up', function() {
     var putExtra = new qiniu.form_up.PutExtra();
 
     // eslint-disable-next-line no-undef
-    describe('test form up#putStreamWithoutKey', function() {
+    describe('test form up#putStreamWithoutKey', function () {
         // eslint-disable-next-line no-undef
-        it('test form up#putStreamWithoutKey', function(done) {
+        it('test form up#putStreamWithoutKey', function (done) {
             var key = null;
             var rs = fs.createReadStream(testFilePath1);
             formUploader.putStream(uploadToken, key, rs, putExtra,
-                function(respErr, respBody, respInfo) {
+                function (respErr, respBody, respInfo) {
                     console.log(respBody, respInfo);
                     should.not.exist(respErr);
                     respBody.should.have.keys('key', 'hash');
@@ -80,13 +78,13 @@ describe('test form up', function() {
     });
 
     // eslint-disable-next-line no-undef
-    describe('test form up#putStream', function() {
+    describe('test form up#putStream', function () {
         // eslint-disable-next-line no-undef
-        it('test form up#putStream', function(done) {
+        it('test form up#putStream', function (done) {
             var key = 'storage_putStream_test' + Math.random(1000);
             var rs = fs.createReadStream(testFilePath1);
             formUploader.putStream(uploadToken, key, rs, putExtra,
-                function(respErr, respBody, respInfo) {
+                function (respErr, respBody, respInfo) {
                     console.log(respBody, respInfo);
                     should.not.exist(respErr);
                     respBody.should.have.keys('key', 'hash');
@@ -97,14 +95,14 @@ describe('test form up', function() {
     });
 
     // eslint-disable-next-line no-undef
-    describe('test form up#put', function() {
+    describe('test form up#put', function () {
         // eslint-disable-next-line no-undef
-        it('test form up#put', function(done) {
+        it('test form up#put', function (done) {
             var key = 'storage_put_test' + Math.random(1000);
             formUploader.put(uploadToken, key, 'hello world', putExtra,
-                function(respErr,
+                function (respErr,
                     respBody) {
-                    //console.log(respBody);
+                    // console.log(respBody);
                     should.not.exist(respErr);
                     respBody.should.have.keys('key', 'hash');
                     keysToDelete.push(respBody.key);
@@ -114,14 +112,14 @@ describe('test form up', function() {
     });
 
     // eslint-disable-next-line no-undef
-    describe('test form up#putWithoutKey', function() {
+    describe('test form up#putWithoutKey', function () {
         // eslint-disable-next-line no-undef
-        it('test form up#putWithoutKey', function(done) {
+        it('test form up#putWithoutKey', function (done) {
             formUploader.putWithoutKey(uploadToken, 'hello world',
                 putExtra,
-                function(respErr,
+                function (respErr,
                     respBody) {
-                    //console.log(respBody);
+                    // console.log(respBody);
                     should.not.exist(respErr);
                     respBody.should.have.keys('key', 'hash');
                     keysToDelete.push(respBody.key);
@@ -131,16 +129,16 @@ describe('test form up', function() {
     });
 
     // eslint-disable-next-line no-undef
-    describe('test form up#putFile', function() {
+    describe('test form up#putFile', function () {
         // eslint-disable-next-line no-undef
-        it('test form up#putFile', function(done) {
+        it('test form up#putFile', function (done) {
             var key = 'storage_putFile_test' + Math.random(1000);
             formUploader.putFile(uploadToken, key, testFilePath2,
                 putExtra,
-                function(
+                function (
                     respErr,
                     respBody) {
-                    //console.log(respBody);
+                    // console.log(respBody);
                     should.not.exist(respErr);
                     respBody.should.have.keys('key', 'hash');
                     keysToDelete.push(respBody.key);
@@ -150,15 +148,15 @@ describe('test form up', function() {
     });
 
     // eslint-disable-next-line no-undef
-    describe('test form up#putFileWithoutKey', function() {
+    describe('test form up#putFileWithoutKey', function () {
         // eslint-disable-next-line no-undef
-        it('test form up#putFileWithoutKey', function(done) {
+        it('test form up#putFileWithoutKey', function (done) {
             formUploader.putFileWithoutKey(uploadToken, testFilePath2,
                 putExtra,
-                function(
+                function (
                     respErr,
                     respBody) {
-                    //console.log(respBody);
+                    // console.log(respBody);
                     should.not.exist(respErr);
                     respBody.should.have.keys('key', 'hash');
                     keysToDelete.push(respBody.key);
