@@ -5,9 +5,8 @@ const proc = require('process');
 const console = require('console');
 
 // eslint-disable-next-line no-undef
-before(function(done) {
-    if (!process.env.QINIU_ACCESS_KEY || !process.env.QINIU_SECRET_KEY || !
-        process.env.QINIU_TEST_BUCKET || !process.env.QINIU_TEST_DOMAIN) {
+before(function (done) {
+    if (!process.env.QINIU_ACCESS_KEY || !process.env.QINIU_SECRET_KEY || !process.env.QINIU_TEST_BUCKET || !process.env.QINIU_TEST_DOMAIN) {
         console.log('should run command `source test-env.sh` first\n');
         process.exit(0);
     }
@@ -15,23 +14,23 @@ before(function(done) {
 });
 
 // eslint-disable-next-line no-undef
-describe('test start bucket manager', function() {
+describe('test start bucket manager', function () {
     var accessKey = proc.env.QINIU_ACCESS_KEY;
     var secretKey = proc.env.QINIU_SECRET_KEY;
     var srcBucket = proc.env.QINIU_TEST_BUCKET;
     var mac = new qiniu.auth.digest.Mac(accessKey, secretKey);
     var config = new qiniu.conf.Config();
-    //config.useHttpsDomain = true;
+    // config.useHttpsDomain = true;
     config.zone = qiniu.zone.Zone_z0;
     var bucketManager = new qiniu.rs.BucketManager(mac, config);
-    //test stat
+    // test stat
     // eslint-disable-next-line no-undef
-    describe('test stat', function() {
+    describe('test stat', function () {
         // eslint-disable-next-line no-undef
-        it('test stat', function(done) {
+        it('test stat', function (done) {
             var bucket = srcBucket;
             var key = 'qiniu.mp4';
-            bucketManager.stat(bucket, key, function(err, respBody,
+            bucketManager.stat(bucket, key, function (err, respBody,
                 respInfo) {
                 console.log(respBody, respInfo);
                 should.not.exist(err);
@@ -42,52 +41,52 @@ describe('test start bucket manager', function() {
         });
     });
 
-    //test copy and move and delete
+    // test copy and move and delete
     // eslint-disable-next-line no-undef
-    describe('test copy', function() {
+    describe('test copy', function () {
         // eslint-disable-next-line no-undef
-        it('test copy', function(done) {
+        it('test copy', function (done) {
             var destBucket = srcBucket;
             var srcKey = 'qiniu.mp4';
             var destKey = 'qiniu_copy.mp4';
             var options = {
-                force: true,
+                force: true
             };
             bucketManager.copy(srcBucket, srcKey, destBucket, destKey,
                 options,
-                function(err, respBody, respInfo) {
-                    //console.log(respBody);
+                function (err, respBody, respInfo) {
+                    // console.log(respBody);
                     should.not.exist(err);
-                    assert.equal(respInfo.statusCode, 200);
+                    assert.strictEqual(respInfo.statusCode, 200);
                     done();
 
-                    //test move
+                    // test move
                     // eslint-disable-next-line no-undef
-                    describe('test move', function() {
+                    describe('test move', function () {
                         var moveDestKey = 'qiniu_move.mp4';
                         // eslint-disable-next-line no-undef
-                        it('test move', function(done1) {
+                        it('test move', function (done1) {
                             bucketManager.move(destBucket, destKey,
                                 destBucket, moveDestKey, options,
-                                function(err1, ret1, info1) {
+                                function (err1, ret1, info1) {
                                     should.not.exist(err1);
-                                    assert.equal(info1.statusCode, 200);
+                                    assert.strictEqual(info1.statusCode, 200);
                                     done1();
 
-                                    //test delete
+                                    // test delete
                                     // eslint-disable-next-line no-undef
-                                    describe('test delete', function() {
+                                    describe('test delete', function () {
                                         // eslint-disable-next-line no-undef
-                                        it('test delete', function(
+                                        it('test delete', function (
                                             done2) {
                                             bucketManager.delete(
                                                 destBucket,
                                                 moveDestKey,
-                                                function(err2, ret2,
+                                                function (err2, ret2,
                                                     info2) {
                                                     should.not.exist(
                                                         err2);
-                                                    assert.equal(info2.statusCode,
+                                                    assert.strictEqual(info2.statusCode,
                                                         200);
                                                     done2();
                                                 });
@@ -101,14 +100,14 @@ describe('test start bucket manager', function() {
     });
 
     // eslint-disable-next-line no-undef
-    describe('test fetch', function() {
+    describe('test fetch', function () {
         // eslint-disable-next-line no-undef
-        it('test fetch', function(done) {
+        it('test fetch', function (done) {
             var resUrl = 'http://devtools.qiniu.com/qiniu.png';
             var bucket = srcBucket;
             var key = 'qiniu.png';
 
-            bucketManager.fetch(resUrl, bucket, key, function(err,
+            bucketManager.fetch(resUrl, bucket, key, function (err,
                 respBody) {
                 should.not.exist(err);
                 respBody.should.have.keys('hash', 'fsize', 'mimeType',
@@ -119,16 +118,16 @@ describe('test start bucket manager', function() {
     });
 
     // eslint-disable-next-line no-undef
-    describe('test changeMime', function() {
+    describe('test changeMime', function () {
         // eslint-disable-next-line no-undef
-        it('test changeMime', function(done) {
+        it('test changeMime', function (done) {
             var key = 'test_file';
             var bucket = srcBucket;
 
             bucketManager.changeMime(bucket, key, 'text/html',
-                function(err, respBody, respInfo) {
+                function (err, respBody, respInfo) {
                     should.not.exist(err);
-                    assert.equal(respInfo.statusCode, 200);
+                    assert.strictEqual(respInfo.statusCode, 200);
                     done();
                 }
             );
@@ -136,37 +135,37 @@ describe('test start bucket manager', function() {
     });
 
     // eslint-disable-next-line no-undef
-    describe('test changeHeaders', function() {
+    describe('test changeHeaders', function () {
         // eslint-disable-next-line no-undef
-        it('test changeHeaders', function(done) {
+        it('test changeHeaders', function (done) {
             var key = 'test_file';
             var bucket = srcBucket;
 
             bucketManager.changeHeaders(bucket, key, {
-                    'Content-Type': 'text/plain',
-                    'Last-Modified': 'Wed, 21 Oct 2015 07:28:00 GMT',
-                    'x-qn-test-custom-header': '0',
-                },
-                function(err, respBody, respInfo) {
-                    console.log(respInfo);
-                    should.not.exist(err);
-                    assert.equal(respInfo.statusCode, 200);
-                    done();
-                }
+                'Content-Type': 'text/plain',
+                'Last-Modified': 'Wed, 21 Oct 2015 07:28:00 GMT',
+                'x-qn-test-custom-header': '0'
+            },
+            function (err, respBody, respInfo) {
+                console.log(respInfo);
+                should.not.exist(err);
+                assert.strictEqual(respInfo.statusCode, 200);
+                done();
+            }
             );
         });
     });
 
     // eslint-disable-next-line no-undef
-    describe('test bucketinfo', function() {
+    describe('test bucketinfo', function () {
         // eslint-disable-next-line no-undef
-        it('test bucketinfo', function(done) {
+        it('test bucketinfo', function (done) {
             var bucket = srcBucket;
 
-            bucketManager.getBucketInfo(bucket, function(err,
+            bucketManager.getBucketInfo(bucket, function (err,
                 respBody, respInfo) {
                 should.not.exist(err);
-                console.log(JSON.stringify(respBody) + "\n");
+                console.log(JSON.stringify(respBody) + '\n');
                 console.log(JSON.stringify(respInfo));
                 done();
             });
@@ -174,143 +173,141 @@ describe('test start bucket manager', function() {
     });
 
     // eslint-disable-next-line no-undef
-    //空间生命周期
-    describe('test lifeRule', function() {
+    // 空间生命周期
+    describe('test lifeRule', function () {
         var bucket = srcBucket;
-        //add
-        describe('test putLifeRule', function() {
-            it('test putLifeRule', function(done) {
+        // add
+        describe('test putLifeRule', function () {
+            it('test putLifeRule', function (done) {
                 var options = {
-                    name: "hello",
-                    prefix: "test",
+                    name: 'hello',
+                    prefix: 'test'
                 };
-                bucketManager.putBucketLifecycleRule(bucket, options, function(err,
+                bucketManager.putBucketLifecycleRule(bucket, options, function (err,
                     respBody, respInfo) {
                     should.not.exist(err);
-                    console.log(JSON.stringify(respBody) + "\n");
+                    console.log(JSON.stringify(respBody) + '\n');
                     console.log(JSON.stringify(respInfo));
                     done();
                 });
             });
         });
 
-        //delete
-        describe('test deleteLifeRule', function() {
-            it('test deleteLifeRule', function(done) {
-                bucketManager.deleteBucketLifecycleRule(bucket, "hello", function(err,
+        // delete
+        describe('test deleteLifeRule', function () {
+            it('test deleteLifeRule', function (done) {
+                bucketManager.deleteBucketLifecycleRule(bucket, 'hello', function (err,
                     respBody, respInfo) {
                     should.not.exist(err);
-                    console.log(JSON.stringify(respBody) + "\n");
+                    console.log(JSON.stringify(respBody) + '\n');
                     console.log(JSON.stringify(respInfo));
                     done();
                 });
             });
         });
 
-        //update
-        describe('test updateLifeRule', function() {
+        // update
+        describe('test updateLifeRule', function () {
             var options = {
-                name: "hello",
+                name: 'hello',
                 history_to_line_after_days: 10,
                 delete_after_days: 10,
-                to_line_after_days: 8,
+                to_line_after_days: 8
             };
-            it('test updateLifeRule', function(done) {
-                bucketManager.updateBucketLifecycleRule(bucket, options, function(err,
+            it('test updateLifeRule', function (done) {
+                bucketManager.updateBucketLifecycleRule(bucket, options, function (err,
                     respBody, respInfo) {
                     should.not.exist(err);
-                    console.log(JSON.stringify(respBody) + "\n");
+                    console.log(JSON.stringify(respBody) + '\n');
                     console.log(JSON.stringify(respInfo));
                     done();
                 });
             });
         });
 
-        //get
-        describe('test getLifeRule', function() {
-            it('test getLifeRule', function(done) {
-                bucketManager.getBucketLifecycleRule(bucket, function(err,
+        // get
+        describe('test getLifeRule', function () {
+            it('test getLifeRule', function (done) {
+                bucketManager.getBucketLifecycleRule(bucket, function (err,
                     respBody, respInfo) {
                     should.not.exist(err);
-                    console.log(JSON.stringify(respBody) + "\n");
+                    console.log(JSON.stringify(respBody) + '\n');
                     console.log(JSON.stringify(respInfo));
                     done();
                 });
             });
         });
-
     });
 
-    describe('test events', function() {
+    describe('test events', function () {
         var bucket = srcBucket;
-        describe('test addEvents', function() {
-            it('test addEvents', function(done) {
+        describe('test addEvents', function () {
+            it('test addEvents', function (done) {
                 var options = {
-                    name: "event_testt",
-                    event: "mkfile",
-                    callbackURL: "http://node.ijemy.com/qncback",
+                    name: 'event_testt',
+                    event: 'mkfile',
+                    callbackURL: 'http://node.ijemy.com/qncback'
                 };
-                bucketManager.putBucketEvent(bucket, options, function(err, respBody, respInfo) {
+                bucketManager.putBucketEvent(bucket, options, function (err, respBody, respInfo) {
                     should.not.exist(err);
-                    console.log(JSON.stringify(respBody) + "\n");
+                    console.log(JSON.stringify(respBody) + '\n');
                     console.log(JSON.stringify(respInfo));
                     done();
                 });
             });
         });
 
-        describe('test updateEvents', function() {
-            it('test updateEvents', function(done) {
+        describe('test updateEvents', function () {
+            it('test updateEvents', function (done) {
                 var options = {
-                    name: "event_testt",
-                    event: "copy",
-                    callbackURL: "http://node.ijemy.com/qncback",
+                    name: 'event_testt',
+                    event: 'copy',
+                    callbackURL: 'http://node.ijemy.com/qncback'
                 };
-                bucketManager.updateBucketEvent(bucket, options, function(err, respBody, respInfo) {
+                bucketManager.updateBucketEvent(bucket, options, function (err, respBody, respInfo) {
                     should.not.exist(err);
-                    console.log(JSON.stringify(respBody) + "\n");
+                    console.log(JSON.stringify(respBody) + '\n');
                     console.log(JSON.stringify(respInfo));
                     done();
                 });
             });
         });
 
-        describe('test getEvents', function() {
-            it('test getEvents', function(done) {
-                bucketManager.getBucketEvent(bucket, function(err, respBody, respInfo) {
+        describe('test getEvents', function () {
+            it('test getEvents', function (done) {
+                bucketManager.getBucketEvent(bucket, function (err, respBody, respInfo) {
                     should.not.exist(err);
-                    console.log(JSON.stringify(respBody) + "\n");
+                    console.log(JSON.stringify(respBody) + '\n');
                     console.log(JSON.stringify(respInfo));
                     done();
                 });
             });
         });
 
-        describe('test deleteEvents', function() {
-            it('test deleteEvents', function(done) {
-                bucketManager.deleteBucketEvent(bucket, "event_test", function(err, respBody, respInfo) {
+        describe('test deleteEvents', function () {
+            it('test deleteEvents', function (done) {
+                bucketManager.deleteBucketEvent(bucket, 'event_test', function (err, respBody, respInfo) {
                     should.not.exist(err);
-                    console.log(JSON.stringify(respBody) + "\n");
+                    console.log(JSON.stringify(respBody) + '\n');
                     console.log(JSON.stringify(respInfo));
                     done();
                 });
             });
         });
-
     });
 
-    describe('test referAntiLeech', function() {
-        describe('test referAntiLeech', function() {
+    describe('test referAntiLeech', function () {
+        describe('test referAntiLeech', function () {
             var options = {
                 mode: 1,
                 norefer: 0,
-                pattern: '*.iorange.vip',
+                pattern: '*.iorange.vip'
             };
             var bucket = srcBucket;
-            it('test referAntiLeech', function(done) {
-                bucketManager.putReferAntiLeech(bucket, options, function(err, respBody, respInfo) {
+            it('test referAntiLeech', function (done) {
+                bucketManager.putReferAntiLeech(bucket, options, function (err, respBody, respInfo) {
                     should.not.exist(err);
-                    console.log(JSON.stringify(respBody) + "\n");
+                    console.log(JSON.stringify(respBody) + '\n');
                     console.log(JSON.stringify(respInfo));
                     done();
                 });
@@ -318,39 +315,39 @@ describe('test start bucket manager', function() {
         });
     });
 
-    describe('test corsRules', function() {
+    describe('test corsRules', function () {
         var bucket = srcBucket;
-        describe('test putCorsRules', function() {
-            it('test putCorsRules', function(done) {
-                var body = new Array();
+        describe('test putCorsRules', function () {
+            it('test putCorsRules', function (done) {
+                var body = [];
                 var req01 = {
-                    "allowed_origin": ["http://www.test1.com"],
-                    "allowed_method": ["GET", "POST"],
+                    allowed_origin: ['http://www.test1.com'],
+                    allowed_method: ['GET', 'POST']
                 };
                 var req02 = {
-                    "allowed_origin": ["http://www.test2.com"],
-                    "allowed_method": ["GET", "POST", "HEAD"],
-                    "allowed_header": ["testheader", "Content-Type"],
-                    "exposed_header": ["test1", "test2"],
-                    "max_age": 20,
-                }
+                    allowed_origin: ['http://www.test2.com'],
+                    allowed_method: ['GET', 'POST', 'HEAD'],
+                    allowed_header: ['testheader', 'Content-Type'],
+                    exposed_header: ['test1', 'test2'],
+                    max_age: 20
+                };
                 body[0] = req01;
                 body[1] = req02;
 
-                bucketManager.putCorsRules(bucket, body, function(err, respBody, respInfo) {
+                bucketManager.putCorsRules(bucket, body, function (err, respBody, respInfo) {
                     should.not.exist(err);
-                    console.log(JSON.stringify(respBody) + "\n");
+                    console.log(JSON.stringify(respBody) + '\n');
                     console.log(JSON.stringify(respInfo));
                     done();
                 });
             });
         });
 
-        describe('test getCorsRules', function() {
-            it('test getCorsRules', function(done) {
-                bucketManager.getCorsRules(bucket, function(err, respBody, respInfo) {
+        describe('test getCorsRules', function () {
+            it('test getCorsRules', function (done) {
+                bucketManager.getCorsRules(bucket, function (err, respBody, respInfo) {
                     should.not.exist(err);
-                    console.log(JSON.stringify(respBody) + "\n");
+                    console.log(JSON.stringify(respBody) + '\n');
                     console.log(JSON.stringify(respInfo));
                     done();
                 });
@@ -375,71 +372,70 @@ describe('test start bucket manager', function() {
     //     });
     // });
 
-    describe('test accessMode', function() {
+    describe('test accessMode', function () {
         var bucket = srcBucket;
-        it('test accessMode', function(done) {
+        it('test accessMode', function (done) {
             var mode = 0;
-            bucketManager.putBucketAccessStyleMode(bucket, mode, function(err, respBody, respInfo) {
+            bucketManager.putBucketAccessStyleMode(bucket, mode, function (err, respBody, respInfo) {
                 should.not.exist(err);
-                console.log(JSON.stringify(respBody) + "\n");
+                console.log(JSON.stringify(respBody) + '\n');
                 console.log(JSON.stringify(respInfo));
                 done();
             });
         });
     });
 
-
-    describe('test putBucketMaxAge', function() {
+    describe('test putBucketMaxAge', function () {
         var bucket = srcBucket;
-        it('test putBucketMaxAge', function(done) {
+        it('test putBucketMaxAge', function (done) {
             var options = {
-                maxAge: 0,
+                maxAge: 0
             };
-            bucketManager.putBucketMaxAge(bucket, options, function(err, respBody, respInfo) {
+            bucketManager.putBucketMaxAge(bucket, options, function (err, respBody, respInfo) {
                 should.not.exist(err);
-                console.log(JSON.stringify(respBody) + "\n");
+                console.log(JSON.stringify(respBody) + '\n');
                 console.log(JSON.stringify(respInfo));
                 done();
             });
         });
     });
 
-    describe('test putBucketAccessMode', function() {
+    describe('test putBucketAccessMode', function () {
         var bucket = srcBucket;
-        it('test putBucketAccessMode', function(done) {
+        it('test putBucketAccessMode', function (done) {
             var options = {
-                private: 0,
+                private: 0
             };
-            bucketManager.putBucketAccessMode(bucket, options, function(err, respBody, respInfo) {
+            bucketManager.putBucketAccessMode(bucket, options, function (err, respBody, respInfo) {
                 should.not.exist(err);
-                console.log(JSON.stringify(respBody) + "\n");
+                console.log(JSON.stringify(respBody) + '\n');
                 console.log(JSON.stringify(respInfo));
                 done();
             });
         });
     });
 
-    describe('test bucketQuota', function() {
+    describe('test bucketQuota', function () {
         var bucket = srcBucket;
-        describe('test putBucketQuota', function() {
-            it('test putBucketQuota', function(done) {
+        describe('test putBucketQuota', function () {
+            it('test putBucketQuota', function (done) {
                 var options = {
                     size: 10,
-                    count: 10,
+                    count: 10
                 };
-                bucketManager.putBucketQuota(bucket, options, function(err, respBody, respInfo) {
+                bucketManager.putBucketQuota(bucket, options, function (err, respBody, respInfo) {
                     should.not.exist(err);
-                    console.log(JSON.stringify(respBody) + "\n");
+                    console.log(JSON.stringify(respBody) + '\n');
                     console.log(JSON.stringify(respInfo));
                     done();
                 });
             });
         });
-        describe('test getBucketQuota', function() {
-            it('test getBucketQuota', function(done) {
-                bucketManager.getBucketQuota(bucket,  function(err, respBody, respInfo) {
+        describe('test getBucketQuota', function () {
+            it('test getBucketQuota', function (done) {
+                bucketManager.getBucketQuota(bucket, function (err, respBody, respInfo) {
                     should.not.exist(err);
-                    console.log(JSON.stringify(respBody) + "\n");
+                    console.log(JSON.stringify(respBody) + '\n');
                     console.log(JSON.stringify(respInfo));
                     done();
                 });
@@ -447,16 +443,15 @@ describe('test start bucket manager', function() {
         });
     });
 
-    describe('test listBucketDomains', function() {
+    describe('test listBucketDomains', function () {
         var bucket = srcBucket;
-        it('test listBucketDomains', function(done) {
-            bucketManager.listBucketDomains(bucket, function(err, respBody, respInfo) {
+        it('test listBucketDomains', function (done) {
+            bucketManager.listBucketDomains(bucket, function (err, respBody, respInfo) {
                 should.not.exist(err);
-                console.log(JSON.stringify(respBody) + "\n");
+                console.log(JSON.stringify(respBody) + '\n');
                 console.log(JSON.stringify(respInfo));
                 done();
             });
         });
     });
-
 });
