@@ -5,7 +5,7 @@ const proc = require('process');
 const console = require('console');
 
 // eslint-disable-next-line no-undef
-before(function(done) {
+before(function (done) {
     if (!process.env.QINIU_PILI_ACCESS_KEY || !process.env.QINIU_PILI_SECRET_KEY) {
         console.log('should run command `source test-env.sh` first\n');
         process.exit(0);
@@ -14,7 +14,7 @@ before(function(done) {
 });
 
 // eslint-disable-next-line no-undef
-describe('test rtc credentials', function() {
+describe('test rtc credentials', function () {
     var accessKey = proc.env.QINIU_ACCESS_KEY;
     var secretKey = proc.env.QINIU_SECRET_KEY;
 
@@ -24,16 +24,16 @@ describe('test rtc credentials', function() {
         hub: 'sdk-live',
         title: 'testtitle',
         maxUsers: 10,
-        noAutoKickUser: true,
+        noAutoKickUser: true
     };
     // eslint-disable-next-line no-undef
-    describe('test create app', function() {
+    describe('test create app', function () {
         // eslint-disable-next-line no-undef
-        it('create app', function(done) {
-            qiniu.app.createApp(appData, credentials, function(err, res) {
+        it('create app', function (done) {
+            qiniu.app.createApp(appData, credentials, function (err, res) {
                 should.not.exist(err);
                 should.exist(res.appId);
-                assert.equal(res.title, 'testtitle');
+                assert.strictEqual(res.title, 'testtitle');
                 appId = res.appId;
                 done();
             });
@@ -41,37 +41,37 @@ describe('test rtc credentials', function() {
     });
 
     // eslint-disable-next-line no-undef
-    describe('test update app', function() {
+    describe('test update app', function () {
         // eslint-disable-next-line no-undef
-        it('update app', function(done) {
+        it('update app', function (done) {
             appData.title = 'testtitle2';
-            qiniu.app.updateApp(appId, appData, credentials, function(err, res) {
+            qiniu.app.updateApp(appId, appData, credentials, function (err, res) {
                 should.not.exist(err);
-                assert.equal(res.title, 'testtitle2');
-                assert.equal(res.appId, appId);
+                assert.strictEqual(res.title, 'testtitle2');
+                assert.strictEqual(res.appId, appId);
                 done();
             });
         });
     });
 
     // eslint-disable-next-line no-undef
-    describe('test get app', function() {
+    describe('test get app', function () {
         // eslint-disable-next-line no-undef
-        it('get app', function(done) {
-            qiniu.app.getApp(appId, credentials, function(err, res) {
+        it('get app', function (done) {
+            qiniu.app.getApp(appId, credentials, function (err, res) {
                 should.not.exist(err);
-                assert.equal(res.title, 'testtitle2');
-                assert.equal(res.appId, appId);
+                assert.strictEqual(res.title, 'testtitle2');
+                assert.strictEqual(res.appId, appId);
                 done();
             });
         });
     });
 
     // eslint-disable-next-line no-undef
-    describe('test delete app', function() {
+    describe('test delete app', function () {
         // eslint-disable-next-line no-undef
-        it('delete app', function(done) {
-            qiniu.app.deleteApp(appId, credentials, function(err) {
+        it('delete app', function (done) {
+            qiniu.app.deleteApp(appId, credentials, function (err) {
                 should.not.exist(err);
                 done();
             });
